# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "File Nest API",
    "version": "1.0.0",
    "description": "RESTful API for file upload, version management, and shareable public links.\n\nResource hierarchy:\n  /files                                    (files)\n    /files/{fileId}\n      /files/{fileId}/versions              (versions of the file)\n        /files/{fileId}/versions/{versionId}\n          /files/{fileId}/versions/{versionId}/content   (binary payload)\n      /files/{fileId}/shares                (share links)\n  /shares/{token}                           (public access via link)"
  },
  "servers": [
    {
      "url": "https://api.filenest.app/v1",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000/v1",
      "description": "Development"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/auth/users": {
      "post": {
        "tags": ["auth"],
        "summary": "Register a new user and start a session",
        "operationId": "registerUser",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Credentials" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created and first session issued",
            "headers": {
              "Location": {
                "schema": { "type": "string", "format": "uri" },
                "description": "URL of the new user resource"
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": {
            "description": "Email already registered",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        }
      }
    },
    "/auth/sessions": {
      "post": {
        "tags": ["auth"],
        "summary": "Create a new session (log in)",
        "description": "Exchanges credentials for a bearer token. The response includes a `Location` header pointing to the new session resource.",
        "operationId": "createSession",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Credentials" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Session created",
            "headers": {
              "Location": {
                "schema": { "type": "string", "format": "uri" },
                "description": "URL of the new session"
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "get": {
        "tags": ["auth"],
        "summary": "List the current user's active sessions",
        "operationId": "listSessions",
        "responses": {
          "200": {
            "description": "Active sessions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Session" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/auth/sessions/current": {
      "get": {
        "tags": ["auth"],
        "summary": "Inspect the caller's current session",
        "operationId": "getCurrentSession",
        "responses": {
          "200": {
            "description": "Current session",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Session" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "delete": {
        "tags": ["auth"],
        "summary": "Revoke the caller's current session (log out)",
        "operationId": "deleteCurrentSession",
        "responses": {
          "204": { "description": "Session revoked" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/auth/sessions/{sessionId}": {
      "parameters": [
        {
          "in": "path",
          "name": "sessionId",
          "required": true,
          "schema": { "type": "string", "format": "uuid" }
        }
      ],
      "delete": {
        "tags": ["auth"],
        "summary": "Revoke a specific session (e.g. sign out another device)",
        "operationId": "deleteSession",
        "responses": {
          "204": { "description": "Session revoked" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/files": {
      "get": {
        "tags": ["files"],
        "summary": "List the current user's files",
        "operationId": "listFiles",
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PerPage" },
          {
            "in": "query",
            "name": "name",
            "schema": { "type": "string" },
            "description": "Filter by name (partial match)"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string",
              "enum": ["createdAt", "-createdAt", "updatedAt", "-updatedAt", "name", "-name"],
              "default": "-updatedAt"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of files",
            "headers": {
              "X-Total-Count": {
                "schema": { "type": "integer" },
                "description": "Total number of matching records"
              },
              "Link": {
                "schema": { "type": "string" },
                "description": "RFC 5988 pagination navigation (next, prev, last)"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/File" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["files"],
        "summary": "Upload a new file (creates the first version)",
        "description": "Creates a new file resource and its initial version (v1). The response includes a `Location` header pointing to the created file.",
        "operationId": "createFile",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["content"],
                "properties": {
                  "content": {
                    "type": "string",
                    "format": "binary",
                    "description": "File binary content"
                  },
                  "name": {
                    "type": "string",
                    "description": "If omitted, the original filename is used"
                  },
                  "description": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "File created",
            "headers": {
              "Location": {
                "schema": { "type": "string", "format": "uri" },
                "description": "URL of the created file"
              },
              "ETag": { "schema": { "type": "string" } }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/File" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "413": { "$ref": "#/components/responses/PayloadTooLarge" }
        }
      }
    },
    "/files/{fileId}": {
      "parameters": [
        { "$ref": "#/components/parameters/FileId" }
      ],
      "get": {
        "tags": ["files"],
        "summary": "Get file metadata",
        "operationId": "getFile",
        "responses": {
          "200": {
            "description": "File metadata",
            "headers": {
              "ETag": { "schema": { "type": "string" } }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/File" }
              }
            }
          },
          "304": { "description": "Not Modified" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["files"],
        "summary": "Update file metadata (partial)",
        "description": "Updates metadata only (name, description). To change content, create a new version.",
        "operationId": "updateFile",
        "parameters": [
          { "$ref": "#/components/parameters/IfMatch" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/merge-patch+json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": ["string", "null"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/File" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "412": { "$ref": "#/components/responses/PreconditionFailed" }
        }
      },
      "delete": {
        "tags": ["files"],
        "summary": "Delete a file and all of its versions",
        "operationId": "deleteFile",
        "responses": {
          "204": { "description": "Deleted" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "FileId": {
        "in": "path",
        "name": "fileId",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
...
      "Page": {
        "in": "query",
        "name": "page",
        "schema": { "type": "integer", "minimum": 1, "default": 1 }
      },
      "PerPage": {
        "in": "query",
        "name": "perPage",
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
      },
      "IfMatch": {
        "in": "header",
        "name": "If-Match",
        "required": false,
        "schema": { "type": "string" },
        "description": "ETag for optimistic concurrency (prevents race conditions)"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid authentication",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      },
      "PreconditionFailed": {
        "description": "If-Match does not match the current ETag",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      },
      "PayloadTooLarge": {
        "description": "File is too large",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      }
    },
    "schemas": {
      "File": {
        "type": "object",
        "required": ["id", "name", "currentVersionId", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": ["string", "null"] },
          "mimeType": { "type": "string" },
          "sizeBytes": { "type": "integer", "format": "int64" },
          "currentVersionId": { "type": "string", "format": "uuid" },
          "versionCount": { "type": "integer" },
          "ownerId": { "type": "string", "format": "uuid" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "_links": {
            "type": "object",
            "description": "HATEOAS navigation",
            "properties": {
              "self": { "type": "string", "format": "uri" },
              "versions": { "type": "string", "format": "uri" },
              "shares": { "type": "string", "format": "uri" },
              "currentContent": { "type": "string", "format": "uri" }
            }
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "RFC 7807 Problem Details",
        "properties": {
          "type": { "type": "string", "format": "uri", "default": "about:blank" },
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "detail": { "type": "string" },
          "instance": { "type": "string", "format": "uri" }
        }
      },
      "Credentials": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 8 }
        }
      },
      "AuthResponse": {
        "type": "object",
        "required": ["token", "expiresAt", "user"],
        "properties": {
          "token": {
            "type": "string",
            "description": "Bearer JWT — send as `Authorization: Bearer <token>`"
          },
          "expiresAt": { "type": "string", "format": "date-time" },
          "session": {
            "type": "object",
            "properties": {
              "id": { "type": "string", "format": "uuid" }
            }
          },
          "user": {
            "type": "object",
            "required": ["id", "email"],
            "properties": {
              "id": { "type": "string", "format": "uuid" },
              "email": { "type": "string", "format": "email" }
            }
          }
        }
      },
      "Session": {
        "type": "object",
        "required": ["id", "createdAt", "lastUsedAt", "expiresAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "userAgent": { "type": ["string", "null"] },
          "ipAddress": { "type": ["string", "null"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "lastUsedAt": { "type": "string", "format": "date-time" },
          "expiresAt": { "type": "string", "format": "date-time" },
          "revokedAt": { "type": ["string", "null"], "format": "date-time" }
        }
      }
    }
  }
}